backend/requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
pydantic
faiss-cpu
numpy
sentence-transformers # For embedding AND re-ranking

# --- LIBRARIES FOR TRANSFORMERS ---
torch
transformers
accelerate           # <-- This was the missing piece